fix: refacto testing functions
