+++ crates/sozo/ops/src/execute.rs
@@ -5,8 +5,7 @@ use starknet::accounts::{Call, ConnectedAccount};
 use starknet::core::types::FieldElement;
 use starknet::core::utils::get_selector_from_name;
 
-use super::get_contract_address;
-use crate::utils::handle_transaction_result;
+use crate::utils;
 
 pub async fn execute<A>(
     contract: String,
@@ -18,7 +17,7 @@ pub async fn execute<A>(
 where
     A: ConnectedAccount + Sync + Send + 'static,
 {
-    let contract_address = get_contract_address(world, contract).await?;
+    let contract_address = utils::get_contract_address(world, contract).await?;
     let res = world
         .account
         .execute(vec![Call {
@@ -30,6 +29,11 @@ where
         .await
         .with_context(|| "Failed to send transaction")?;
 
-    handle_transaction_result(&world.account.provider(), res, transaction.wait, transaction.receipt)
-        .await
+    utils::handle_transaction_result(
+        &world.account.provider(),
+        res,
+        transaction.wait,
+        transaction.receipt,
+    )
+    .await
 }
