+++ crates/sozo/ops/src/tests/auth.rs
@@ -1,62 +1,23 @@
-use anyhow::Result;
-use dojo_test_utils::compiler::build_test_config;
-use dojo_test_utils::migration::prepare_migration;
 use dojo_test_utils::sequencer::{
     get_default_test_starknet_config, SequencerConfig, TestSequencer,
 };
 use dojo_world::contracts::world::WorldContract;
 use dojo_world::migration::TxConfig;
-use scarb::ops;
-use starknet::accounts::{Account, ConnectedAccount, SingleOwnerAccount};
-use starknet::core::types::{BlockId, BlockTag};
+use starknet::accounts::{Account, ConnectedAccount};
 use starknet::core::utils::cairo_short_string_to_felt;
-use starknet::providers::jsonrpc::HttpTransport;
-use starknet::providers::JsonRpcClient;
-use starknet::signers::LocalWallet;
 
+use super::setup;
 use crate::auth::{self, ModelContract, OwnerResource, ResourceType};
-use crate::{execute, migration};
+use crate::execute;
 
 const ACTION_CONTRACT_NAME: &str = "dojo_examples::actions::actions";
-
-/// Setups the project by migrating the spawn-and-moves project.
-///
-/// # Returns
-///
-/// A [`WorldContract`] initialized with the migrator account,
-/// the account 0 of the sequencer.
-async fn setup(
-    sequencer: &TestSequencer,
-) -> Result<WorldContract<SingleOwnerAccount<JsonRpcClient<HttpTransport>, LocalWallet>>> {
-    let config = build_test_config("../../../examples/spawn-and-move/Scarb.toml")?;
-    let ws = ops::read_workspace(config.manifest_path(), &config)
-        .unwrap_or_else(|op| panic!("Error building workspace: {op:?}"));
-    let base_dir = "../../../examples/spawn-and-move";
-    let target_dir = format!("{}/target/dev", base_dir);
-
-    let migration = prepare_migration(base_dir.into(), target_dir.into())?;
-
-    let mut account = sequencer.account();
-    account.set_block_id(BlockId::Tag(BlockTag::Pending));
-
-    let output = migration::execute_strategy(
-        &ws,
-        &migration,
-        &account,
-        Some(TxConfig { wait: true, ..Default::default() }),
-    )
-    .await?;
-    let world = WorldContract::new(output.world_address, account);
-
-    Ok(world)
-}
 
 #[tokio::test(flavor = "multi_thread")]
 async fn auth_grant_writer_ok() {
     let sequencer =
         TestSequencer::start(SequencerConfig::default(), get_default_test_starknet_config()).await;
 
-    let world = setup(&sequencer).await.unwrap();
+    let world = setup::setup(&sequencer).await.unwrap();
 
     // Shouldn't have any permission at this point.
     let account2 = sequencer.account_at_index(2);
@@ -95,7 +56,7 @@ async fn auth_revoke_writer_ok() {
     let sequencer =
         TestSequencer::start(SequencerConfig::default(), get_default_test_starknet_config()).await;
 
-    let world = setup(&sequencer).await.unwrap();
+    let world = setup::setup(&sequencer).await.unwrap();
 
     // Shouldn't have any permission at this point.
     let account2 = sequencer.account_at_index(2);
@@ -144,7 +105,7 @@ async fn auth_grant_owner_ok() {
     let sequencer =
         TestSequencer::start(SequencerConfig::default(), get_default_test_starknet_config()).await;
 
-    let world = setup(&sequencer).await.unwrap();
+    let world = setup::setup(&sequencer).await.unwrap();
 
     // Shouldn't have any permission at this point.
     let account_2 = sequencer.account_at_index(2);
@@ -179,7 +140,7 @@ async fn auth_revoke_owner_ok() {
     let sequencer =
         TestSequencer::start(SequencerConfig::default(), get_default_test_starknet_config()).await;
 
-    let world = setup(&sequencer).await.unwrap();
+    let world = setup::setup(&sequencer).await.unwrap();
 
     // Shouldn't have any permission at this point.
     let account_2 = sequencer.account_at_index(2);
