+++ crates/sozo/ops/src/tests/mod.rs
@@ -1,1 +1,3 @@
 pub mod auth;
+pub mod setup;
+pub mod utils;
